+++ R/read_vcf.R
@@ -1,5 +1,8 @@
 #' Read VCF Files as MAF Object
 #'
+#' MAF file is more recommended. In this function, we will mimic
+#' the MAF object from the key `c(1, 2, 4, 5, 7)` columns of VCF file.
+#'
 #' @param vcfs VCF file paths.
 #' @param samples sample names for VCF files.
 #' @param genome_build genome build version like "hg19".
@@ -13,15 +16,21 @@
 #' vcfs <- list.files(system.file("extdata", package = "sigminer"), "*.vcf", full.names = TRUE)
 #' \donttest{
 #' maf <- read_vcf(vcfs)
-#' maf <- read_vcf(vcfs, keep_only_pass = FALSE)
+#' maf <- read_vcf(vcfs, keep_only_pass = TRUE)
 #' }
 #' @testexamples
 #' expect_is(maf, "MAF")
-read_vcf <- function(vcfs, samples = NULL, genome_build = c("hg19", "hg38", "mm10"), keep_only_pass = TRUE, verbose = TRUE) {
+read_vcf <- function(vcfs, samples = NULL, genome_build = c("hg19", "hg38", "mm10"), keep_only_pass = FALSE, verbose = TRUE) {
   genome_build <- match.arg(genome_build)
   vcfs_name <- vcfs
   if (verbose) message("Reading file(s): ", paste(vcfs, collapse = ", "))
-  vcfs <- purrr::map(vcfs, ~ data.table::fread(., select = c(1, 2, 4, 5, 7)))
+  vcfs <- purrr::map(vcfs, ~ tryCatch(
+    data.table::fread(., select = c(1, 2, 4, 5, 7), skip = "#CHROM"),
+    error = function(e) {
+      message("It seems ", ., " has no normal VCF header, try parsing without header.")
+      data.table::fread(., select = c(1, 2, 4, 5, 7))
+    }
+  ))
 
   if (is.null(samples)) {
     names(vcfs) <- file_name(vcfs_name, must_chop = ".vcf")